+++ src/app/service/poap/DistributePOAP.ts
@@ -5,12 +5,12 @@ import ValidationError from '../../errors/ValidationError';
 import { Db } from 'mongodb';
 import constants from '../constants/constants';
 import { CommandContext } from 'slash-create';
-import { LogUtils } from '../../utils/Log';
+import Log, { LogUtils } from '../../utils/Log';
 import { Buffer } from 'buffer';
 import { getBufferForFailedParticipants } from './EndPOAP';
 import MongoDbUtils from '../../utils/dbUtils';
 
-export default async (ctx: CommandContext, guildMember: GuildMember, event?: string): Promise<any> => {
+export default async (ctx: CommandContext, guildMember: GuildMember, event: string, code?: string): Promise<any> => {
 	const db: Db = await MongoDbUtils.connect(constants.DB_NAME_DEGEN);
 	await POAPUtils.validateUserAccess(guildMember, db);
 	await POAPUtils.validateEvent(guildMember, event);
@@ -33,10 +33,15 @@ export default async (ctx: CommandContext, guildMember: GuildMember, event?: str
 		],
 		files: [{ name: 'failed_to_send_poaps.csv', attachment: failedPOAPsBuffer }],
 	});
+	if (failedPOAPsList.length <= 0) {
+		Log.debug('all poap successfully delivered');
+		return;
+	}
+	await POAPUtils.setupFailedAttendeesDelivery(guildMember, failedPOAPsList, event, code, ctx);
 };
 
 export const askForParticipantsList = async (guildMember: GuildMember): Promise<POAPFileParticipant[]> => {
-	const message: Message = await guildMember.send({ content: 'Please upload participants.csv file.' });
+	const message: Message = await guildMember.send({ content: 'Please upload participants.csv file. POAPs will be distributed to these degens.' });
 	const dmChannel: DMChannel = await message.channel.fetch() as DMChannel;
 	const replyOptions: AwaitMessagesOptions = {
 		max: 1,